+++ src/collection.rs
@@ -1,7 +1,8 @@
 use core::marker::PhantomData;
 
 use embedded_graphics::{
-    prelude::{DrawTarget, PixelColor, Point, Size},
+    pixelcolor::BinaryColor,
+    prelude::{DrawTarget, Point, Size},
     primitives::Rectangle,
 };
 use embedded_layout::{object_chain::ChainElement, prelude::*, view_group::ViewGroup};
@@ -9,6 +10,7 @@ use embedded_layout::{object_chain::ChainElement, prelude::*, view_group::ViewGr
 use crate::{
     interaction::InputAdapterSource,
     selection_indicator::{style::IndicatorStyle, SelectionIndicatorController},
+    theme::Theme,
     Marker, MenuItem, MenuStyle,
 };
 
@@ -29,8 +31,8 @@ pub trait MenuItemCollection<R> {
         S: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        D: DrawTarget<Color = C>,
-        C: PixelColor + Default + 'static;
+        D: DrawTarget<Color = BinaryColor>,
+        C: Theme;
 }
 
 // Treat any MenuItem impl as a 1-element collection
@@ -71,8 +73,8 @@ where
         S: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        DIS: DrawTarget<Color = C>,
-        C: PixelColor + Default + 'static,
+        DIS: DrawTarget<Color = BinaryColor>,
+        C: Theme,
     {
         MenuItem::draw_styled(self, style, display)
     }
@@ -135,17 +137,17 @@ where
         self.items.as_ref().len()
     }
 
-    fn draw_styled<MSC, S, IT, P, D>(
+    fn draw_styled<T, S, IT, P, D>(
         &self,
-        style: &MenuStyle<S, IT, P, R, MSC>,
+        style: &MenuStyle<S, IT, P, R, T>,
         display: &mut D,
     ) -> Result<(), D::Error>
     where
         S: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        D: DrawTarget<Color = MSC>,
-        MSC: PixelColor + Default + 'static,
+        D: DrawTarget<Color = BinaryColor>,
+        T: Theme,
     {
         for item in self.items.as_ref() {
             item.draw_styled(style, display)?;
@@ -233,8 +235,8 @@ where
         S: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        D: DrawTarget<Color = C>,
-        C: PixelColor + Default + 'static,
+        D: DrawTarget<Color = BinaryColor>,
+        C: Theme,
     {
         self.object.draw_styled(style, display)
     }
@@ -294,8 +296,8 @@ where
         S: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        D: DrawTarget<Color = C>,
-        C: PixelColor + Default + 'static,
+        D: DrawTarget<Color = BinaryColor>,
+        C: Theme,
     {
         self.parent.draw_styled(style, display)?;
         self.object.draw_styled(style, display)?;