+++ src/items/section_title.rs
@@ -1,5 +1,6 @@
 use embedded_graphics::{
-    prelude::{DrawTarget, PixelColor, Point},
+    pixelcolor::BinaryColor,
+    prelude::{DrawTarget, Point},
     primitives::Rectangle,
 };
 use embedded_layout::View;
@@ -8,6 +9,7 @@ use crate::{
     interaction::InputAdapterSource,
     items::MenuLine,
     selection_indicator::{style::IndicatorStyle, SelectionIndicatorController},
+    theme::Theme,
     Marker, MenuItem, MenuStyle,
 };
 
@@ -42,7 +44,7 @@ where
         S: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        C: PixelColor,
+        C: Theme,
     {
         self.line = MenuLine::new("", style);
     }
@@ -56,8 +58,8 @@ where
         S: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        D: DrawTarget<Color = C>,
-        C: PixelColor + Default + 'static,
+        D: DrawTarget<Color = BinaryColor>,
+        C: Theme,
     {
         self.line
             .draw_styled(self.title_text.as_ref(), "", style, display)