+++ src/items/select.rs
@@ -1,5 +1,6 @@
 use embedded_graphics::{
-    prelude::{DrawTarget, PixelColor, Point},
+    pixelcolor::BinaryColor,
+    prelude::{DrawTarget, Point},
     primitives::Rectangle,
 };
 use embedded_layout::View;
@@ -8,6 +9,7 @@ use crate::{
     interaction::InputAdapterSource,
     items::MenuLine,
     selection_indicator::{style::IndicatorStyle, SelectionIndicatorController},
+    theme::Theme,
     Marker, MenuItem, MenuStyle,
 };
 
@@ -93,12 +95,12 @@ where
         (self.convert)(self.value)
     }
 
-    fn set_style<ST, IT, P, C>(&mut self, style: &MenuStyle<ST, IT, P, R, C>)
+    fn set_style<IS, IT, P, C>(&mut self, style: &MenuStyle<IS, IT, P, R, C>)
     where
-        ST: IndicatorStyle,
+        IS: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        C: PixelColor,
+        C: Theme,
     {
         let initial = self.value;
         let mut longest_str = initial.name();
@@ -123,8 +125,8 @@ where
         IS: IndicatorStyle,
         IT: InputAdapterSource<R>,
         P: SelectionIndicatorController,
-        DIS: DrawTarget<Color = C>,
-        C: PixelColor + Default + 'static,
+        DIS: DrawTarget<Color = BinaryColor>,
+        C: Theme,
     {
         self.line
             .draw_styled(self.title_text.as_ref(), self.value.name(), style, display)