+++ src/selection_indicator/style/animated_triangle.rs
@@ -1,9 +1,7 @@
 use embedded_graphics::{
-    pixelcolor::BinaryColor,
-    prelude::{DrawTarget, PixelColor, Point},
+    prelude::{DrawTarget, Point},
     primitives::Rectangle,
     transform::Transform,
-    Drawable,
 };
 
 use crate::{
@@ -12,17 +10,17 @@ use crate::{
         style::{interpolate, triangle::Arrow, IndicatorStyle},
         Insets,
     },
+    theme::Theme,
 };
 
 #[derive(Clone, Copy)]
-pub struct AnimatedTriangle<C = BinaryColor> {
+pub struct AnimatedTriangle {
     period: i32,
-    color: C,
 }
 
-impl<C> AnimatedTriangle<C> {
-    pub const fn new(period: i32, color: C) -> Self {
-        Self { period, color }
+impl AnimatedTriangle {
+    pub const fn new(period: i32) -> Self {
+        Self { period }
     }
 }
 
@@ -31,13 +29,9 @@ pub struct State {
     current: i32,
 }
 
-impl<C> IndicatorStyle for AnimatedTriangle<C>
-where
-    C: Copy + PixelColor,
-{
-    type Shape = Arrow<C>;
+impl IndicatorStyle for AnimatedTriangle {
+    type Shape = Arrow;
     type State = State;
-    type Color = C;
 
     fn on_target_changed(&self, state: &mut Self::State) {
         state.current = 0;
@@ -61,7 +55,7 @@ where
     }
 
     fn shape(&self, state: &Self::State, bounds: Rectangle, fill_width: u32) -> Self::Shape {
-        let max_offset = Arrow::<C>::tip_width(bounds);
+        let max_offset = Self::Shape::tip_width(bounds);
 
         let half_move = self.period / 5;
         let rest = 3 * half_move;
@@ -76,21 +70,19 @@ where
 
         let offset = offset * max_offset / half_move;
 
-        Arrow::new(bounds, fill_width, self.color(state)).translate(Point::new(-offset, 0))
+        Arrow::new(bounds, fill_width).translate(Point::new(-offset, 0))
     }
 
-    fn color(&self, _state: &Self::State) -> Self::Color {
-        self.color
-    }
-
-    fn draw<D>(
+    fn draw<T, D>(
         &self,
         state: &Self::State,
         input_state: InputState,
+        theme: &T,
         display: &mut D,
     ) -> Result<Self::Shape, D::Error>
     where
-        D: DrawTarget<Color = C>,
+        T: Theme,
+        D: DrawTarget<Color = T::Color>,
     {
         let display_area = display.bounding_box();
 
@@ -102,7 +94,7 @@ where
 
         let shape = self.shape(state, display_area, fill_width);
 
-        shape.draw(display)?;
+        shape.draw(theme.selection_color(), display)?;
 
         Ok(shape)
     }