+++ src/selection_indicator/style/mod.rs
@@ -1,21 +1,20 @@
 use embedded_graphics::{
-    prelude::{DrawTarget, PixelColor},
+    prelude::DrawTarget,
     primitives::{ContainsPoint, Rectangle},
     transform::Transform,
 };
 
-use crate::{interaction::InputState, selection_indicator::Insets};
+use crate::{interaction::InputState, selection_indicator::Insets, theme::Theme};
 
 pub mod animated_triangle;
 pub mod border;
-// pub mod invert;
 pub mod line;
+pub mod rectangle;
 pub mod triangle;
 
 // Re-export the styles themselves to make them easier to use.
 pub use animated_triangle::AnimatedTriangle;
 pub use border::Border;
-// pub use invert::Invert;
 pub use line::Line;
 pub use triangle::Triangle;
 
@@ -33,24 +32,24 @@ pub fn interpolate(value: u32, x_min: u32, x_max: u32, y_min: u32, y_max: u32) -
     }
 }
 
-pub trait IndicatorStyle: Clone + Copy {
+pub trait IndicatorStyle: Copy {
     type Shape: ContainsPoint + Transform + Clone;
     type State: Default + Copy;
-    type Color: PixelColor;
 
     fn on_target_changed(&self, _state: &mut Self::State) {}
     fn update(&self, _state: &mut Self::State, _input_state: InputState) {}
     fn padding(&self, state: &Self::State, height: i32) -> Insets;
     fn shape(&self, state: &Self::State, bounds: Rectangle, fill_width: u32) -> Self::Shape;
-    fn color(&self, state: &Self::State) -> Self::Color;
-    fn draw<D>(
+    fn draw<T, D>(
         &self,
         state: &Self::State,
         input_state: InputState,
+        theme: &T,
         display: &mut D,
     ) -> Result<Self::Shape, D::Error>
     where
-        D: DrawTarget<Color = Self::Color>;
+        T: Theme,
+        D: DrawTarget<Color = T::Color>;
 }
 
 #[test]