+++ src/selection_indicator/style/rectangle.rs
@@ -1,10 +1,8 @@
 use embedded_graphics::{
-    prelude::{DrawTarget, Point, Size},
-    primitives::{ContainsPoint, Primitive, PrimitiveStyle, Rectangle as RectangleShape},
-    transform::Transform,
+    prelude::DrawTarget,
+    primitives::{Primitive, PrimitiveStyle, Rectangle as RectangleShape},
     Drawable,
 };
-use embedded_layout::prelude::{horizontal::LeftToRight, vertical::Center, Align};
 
 use crate::{
     interaction::InputState,
@@ -12,41 +10,39 @@ use crate::{
         style::{interpolate, IndicatorStyle},
         Insets,
     },
+    theme::Theme,
 };
 
 #[derive(Clone, Copy)]
-pub struct Rectangle<C = BinaryColor> {
-    color: C,
-}
+pub struct Rectangle;
 
-impl<C> IndicatorStyle for Rectangle<C> 
-where C: Copy
-{
-    type Shape = Arrow<C>;
+impl IndicatorStyle for Rectangle {
+    type Shape = RectangleShape;
     type State = ();
 
-    fn padding(&self, _state: &Self::State, height: i32) -> Insets {
+    fn padding(&self, _state: &Self::State, _height: i32) -> Insets {
         Insets {
-            left: height / 2 + 1,
+            left: 2,
             top: 0,
             right: 0,
             bottom: 0,
         }
     }
 
-    fn shape(&self, _state: &Self::State, bounds: RectangleShape, fill_width: u32) -> Self::Shape {
-        Arrow::new(bounds, fill_width, self.color)
+    fn shape(&self, _state: &Self::State, bounds: RectangleShape, _fill_width: u32) -> Self::Shape {
+        bounds
     }
 
-    fn draw<D>(
+    fn draw<T, D>(
         &self,
         state: &Self::State,
         input_state: InputState,
+        theme: &T,
         display: &mut D,
     ) -> Result<Self::Shape, D::Error>
     where
-        D: DrawTarget,
-        <D as DrawTarget>::Color: From<<Self::Shape as Drawable>::Color>,
+        T: Theme,
+        D: DrawTarget<Color = T::Color>,
     {
         let display_area = display.bounding_box();
 
@@ -58,7 +54,9 @@ where C: Copy
 
         let shape = self.shape(state, display_area, fill_width);
 
-        shape.draw(display)?;
+        shape
+            .into_styled(PrimitiveStyle::with_fill(theme.selection_color()))
+            .draw(display)?;
 
         Ok(shape)
     }