+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Unreleased
  - Removed `derive(Menu)`
  - Removed the `R` type parameter from `SectionTitle`
  - Added initial color support
+ - Style a menu by using a `Theme` object
+ - Replaced the `Invert` selection indicator with `Rectangle`
+ - Removed `IndicatorStyle::Color`, `IndicatorStyle::color`, `IndicatorStyle::draw` is now generic over the theme
 
 0.5.4 (2023-10-27)
 ==================